Add a bit more structure
