+++ Clean swift example/Onboarding/Interactor/OnboardingInteractor.swift
@@ -1,11 +1,5 @@
 import Foundation
 
-protocol OnboardingInteractorProtocol {
-    func completeOnboarding()
-    
-    func showOnboardingView(request: Model.ShowOnboardingView.Request)
-}
-
 final class OnboardingInteractor {
     
     let presenter: OnboardingPresenterProtocol
