+++ Clean swift example/Onboarding/Interactor/OnboardingInteractorProtocol.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+protocol OnboardingInteractorProtocol {
+
+    func completeOnboarding()
+    func showOnboardingView(request: Model.ShowOnboardingView.Request)
+}
