+++ Clean swift example/Onboarding/Presenter/OnboardingPresenterProtocol.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+protocol OnboardingPresenterProtocol {
+ 
+    func dismissOnboarding()
+    func presentOnboarding(response: Model.ShowOnboardingView.Response)
+}
