+++ Clean swift example/Onboarding/View/OnboardingViewController.swift
@@ -1,11 +1,5 @@
 import UIKit
 
-protocol OnboardingViewProtocol: class {
-    func dismiss()
-    
-    func showOnboardingView(viewModel: Model.ShowOnboardingView.ViewModel)
-}
-
 class OnboardingViewController: UIViewController {
     
     @IBOutlet weak var titleLabel: UILabel!
