+++ Clean swift example/Onboarding/View/OnboardingViewProtocol.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+protocol OnboardingViewProtocol: class {
+    
+    func dismiss()
+    func showOnboardingView(viewModel: Model.ShowOnboardingView.ViewModel)
+}
