+++ Clean swift example/Onboarding/Worker/OnboardingWorker.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-final class OnboardingWorker {
+final class OnboardingWorker: OnboardingWorkerProtocol {
     func completeOnboarding() {
         UserDefaults.standard.set(true, forKey: "onboarding-complete")
     }
