+++ Clean swift example/Onboarding/Worker/OnboardingWorkerProtocol.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+protocol OnboardingWorkerProtocol: class {
+
+    func completeOnboarding()
+}
