+++ Clean swift example/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
