+++ Clean swift example/Resources/Assets.xcassets/Contents.json
