+++ Clean swift example/Resources/Base.lproj/LaunchScreen.storyboard
