+++ Clean swift example/Resources/Base.lproj/Main.storyboard
