+++ Clean swift example/Resources/Info.plist
